Fix theme crash and cleanup
